Enhance project configuration and error handling
- Updated .hintrc to include new accessibility hint for keyboard navigation
- Adjusted Jest configuration to lower function coverage threshold for improved testing flexibility
- Modified package.json to streamline the development script by removing the turbo flag
- Restored Sentry error capturing in GlobalError component and removed debugging console.error
- Integrated BodyClassManager component into layout for dynamic body class management based on window states
- Created BodyClassManager component to handle body class updates based on global window states, improving HMR stability
- Refactored GlobalWindowRegistryContext to remove unnecessary state management related to body class handling
