+++ jest.config.js
@@ -15,7 +15,7 @@ const customJestConfig = {
     global: {
       statements: 65,
       branches: 55,
-      functions: 75,
+      functions: 65,
       lines: 65
     }
   },
